requirements.txt
boto3==1.16.25
botocore==1.19.25
click==7.1.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==8.0.1
python-dateutil==2.8.1
requests==2.7.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.2.12
urllib3==1.26.2
Werkzeug==0.14.1
